Abandon all faith in webpack
Trying to use webpack as a compiler is just not going to work. The architecture is fighting us at every step. Instead, we'll just fall back to doing our own project analysis and using webpack as a dumb pipe for getting the data segment into a build as a module. This change fixes all of the incompatibilities caused by using the need-additional-reseal feature.